update week4 folder

+++ Week4/script.py
@@ -19,7 +19,7 @@ def predict():
     '''
     DEFINE PREDICTION FUCTION
     '''
-    int_features = [int(x) for x in request.form.values()]
+    int_features = [float(x) for x in request.form.values()]
     final_features = [np.array(int_features)]
     prediction = model.predict(final_features)
